+++ markets/spot-market/contracts/modules/SpotMarketFactoryModule.sol
@@ -99,8 +99,8 @@ contract SpotMarketFactoryModule is ISpotMarketFactoryModule, AssociatedSystemsM
      * @dev collateral leverage is defaulted to 1 on registration of a new market
      */
     function locked(uint128 marketId) external view returns (uint256) {
-        uint totalBalance = SynthUtil.getToken(marketId).totalSupply();
-        uint collateralLeverage = MarketConfiguration.load(marketId).collateralLeverage;
+        uint256 totalBalance = SynthUtil.getToken(marketId).totalSupply();
+        uint256 collateralLeverage = MarketConfiguration.load(marketId).collateralLeverage;
 
         return
             collateralLeverage == 0
