+++ markets/spot-market/contracts/utils/MathUtil.sol
@@ -9,28 +9,28 @@ library MathUtil {
     using SafeCastU256 for uint256;
     using DecimalMath for int256;
 
-    function abs(int x) internal pure returns (uint) {
+    function abs(int256 x) internal pure returns (uint256) {
         return x >= 0 ? x.toUint() : (-x).toUint();
     }
 
-    function max(int x, int y) internal pure returns (int) {
+    function max(int256 x, int256 y) internal pure returns (int256) {
         return x < y ? y : x;
     }
 
-    function min(int x, int y) internal pure returns (int) {
+    function min(int256 x, int256 y) internal pure returns (int256) {
         return x < y ? x : y;
     }
 
-    function min(uint x, uint y) internal pure returns (uint) {
+    function min(uint256 x, uint256 y) internal pure returns (uint256) {
         return x < y ? x : y;
     }
 
-    function sameSide(int a, int b) internal pure returns (bool) {
+    function sameSide(int256 a, int256 b) internal pure returns (bool) {
         return (a == 0) || (b == 0) || (a > 0) == (b > 0);
     }
 
-    function sqrt(int x) internal pure returns (int y) {
-        int z = (x + 1) / 2;
+    function sqrt(int256 x) internal pure returns (int256 y) {
+        int256 z = (x + 1) / 2;
         y = x;
         while (z < y) {
             y = z;
@@ -38,7 +38,7 @@ library MathUtil {
         }
     }
 
-    function pow(int x, uint n) internal pure returns (int r) {
+    function pow(int256 x, uint256 n) internal pure returns (int256 r) {
         r = DecimalMath.UNIT_INT;
         while (n > 0) {
             if (n % 2 == 1) {
